# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Millitary service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/soldier/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new soldier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soldier"
                ],
                "summary": "Create Soldier",
                "parameters": [
                    {
                        "description": "Soldier Request",
                        "name": "SoldierReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/soldiers.CreateSoldier"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while creating",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/soldier/dashbord": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all Dashbord",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashbord"
                ],
                "summary": "Get All Dashbord",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Join date of the soldier (format: YYYY-MM-DD)",
                        "name": "join_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date of the soldier (format: YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get All Successful",
                        "schema": {
                            "$ref": "#/definitions/soldiers.AllSoldiers"
                        }
                    },
                    "401": {
                        "description": "Error while getting all",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/soldier/delete/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an existing soldier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soldier"
                ],
                "summary": "Delete Soldier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Soldier ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while deleting",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/soldier/get/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get an existing soldier by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soldier"
                ],
                "summary": "Get Soldier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Soldier ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get Successful",
                        "schema": {
                            "$ref": "#/definitions/soldiers.Soldier"
                        }
                    },
                    "401": {
                        "description": "Error while getting",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/soldier/getall": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all soldiers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soldier"
                ],
                "summary": "Get All Soldiers",
                "parameters": [
                    {
                        "type": "string",
                        "name": "age",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get All Successful",
                        "schema": {
                            "$ref": "#/definitions/soldiers.AllSoldiers"
                        }
                    },
                    "401": {
                        "description": "Error while getting all",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/soldier/getallfuelstatistik": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all fuel statistics for soldiers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashbord"
                ],
                "summary": "Get All Fuel Statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in the format YYYY-MM-DD",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Soldier ID",
                        "name": "soldier_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get All Successful",
                        "schema": {
                            "$ref": "#/definitions/soldiers.GetSoldierStatistikFuelRes"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/soldier/getallweaponstatistik": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all weapon statistics for soldiers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dashbord"
                ],
                "summary": "Get All Weapon Statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date in the format YYYY-MM-DD",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Soldier ID",
                        "name": "soldier_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get All Successful",
                        "schema": {
                            "$ref": "#/definitions/soldiers.GetSoldierStatistikRes"
                        }
                    },
                    "400": {
                        "description": "Invalid query parameter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/soldier/update/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing soldier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soldier"
                ],
                "summary": "Update Soldier",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Soldier ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Soldier",
                        "name": "Soldier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/soldiers.Soldier"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while updating",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/soldier/usebullet": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Record the use of bullets by a soldier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soldier"
                ],
                "summary": "Use Bullet",
                "parameters": [
                    {
                        "description": "Use Bullet",
                        "name": "UseB",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/soldiers.UseB"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Use Bullet Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while using bullet",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/soldier/usefuel": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Record the use of fuel by a soldier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Soldier"
                ],
                "summary": "Use Fuel",
                "parameters": [
                    {
                        "description": "Use Fuel",
                        "name": "UseF",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/soldiers.UseF"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Use Fuel Successful",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Error while using fuel",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "soldiers.AllSoldiers": {
            "type": "object",
            "properties": {
                "soldiers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/soldiers.Soldier"
                    }
                }
            }
        },
        "soldiers.Commander": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                }
            }
        },
...
        "soldiers.CreateSoldier": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "group": {
                    "$ref": "#/definitions/soldiers.Group"
                },
                "join_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "soldiers.Department": {
            "type": "object",
            "properties": {
                "commander": {
                    "$ref": "#/definitions/soldiers.Commander"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "soldiers.GetSoldierStatistikFuelRes": {
            "type": "object",
            "properties": {
                "used_fuel": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/soldiers.UseF"
                    }
                }
            }
        },
        "soldiers.GetSoldierStatistikRes": {
            "type": "object",
            "properties": {
                "used_weapons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/soldiers.UseB"
                    }
                }
            }
        },
        "soldiers.Group": {
            "type": "object",
            "properties": {
                "department": {
                    "$ref": "#/definitions/soldiers.Department"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "soldiers.Soldier": {
            "type": "object",
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "group": {
                    "$ref": "#/definitions/soldiers.Group"
                },
                "id": {
                    "type": "string"
                },
                "join_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "soldiers.UseB": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "quantity_big_weapon": {
                    "type": "integer"
                },
                "quantity_weapon": {
                    "type": "integer"
                },
                "soldier_id": {
                    "type": "string"
                }
            }
        },
        "soldiers.UseF": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "diesel": {
                    "type": "integer"
                },
                "petrol": {
                    "type": "integer"
                },
                "soldier_id": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authourization",
            "in": "header"
        }
    }
}